third_party/WebKit/LayoutTests/accessibility/inline-text-word-boundaries-expected.txt

Ace ten twenty-one thirty-five. xxxxxxxxxxxxxxxxxxx

Tests that we can compute word boundaries for any character index from the accessibility tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Accessible text: "Ace ten twenty-one thirty-five. xxxxxxxxxxxxxxxxxxx"
Character 0: 'A' word=(0, 3): 'Ace'
Character 1: 'c' word=(0, 3): 'Ace'
Character 2: 'e' word=(0, 3): 'Ace'
Character 3: ' ' word=(0, 3): 'Ace'
Character 4: 't' word=(4, 7): 'ten'
Character 5: 'e' word=(4, 7): 'ten'
Character 6: 'n' word=(4, 7): 'ten'
Character 7: ' ' word=(4, 7): 'ten'
Character 8: 't' word=(8, 14): 'twenty'
Character 9: 'w' word=(8, 14): 'twenty'
Character 10: 'e' word=(8, 14): 'twenty'
Character 11: 'n' word=(8, 14): 'twenty'
Character 12: 't' word=(8, 14): 'twenty'
Character 13: 'y' word=(8, 14): 'twenty'
Character 14: '-' word=(8, 14): 'twenty'
Character 15: 'o' word=(15, 18): 'one'
Character 16: 'n' word=(15, 18): 'one'
Character 17: 'e' word=(15, 18): 'one'
Character 18: ' ' word=(15, 18): 'one'
Character 19: 't' word=(19, 25): 'thirty'
Character 20: 'h' word=(19, 25): 'thirty'
Character 21: 'i' word=(19, 25): 'thirty'
Character 22: 'r' word=(19, 25): 'thirty'
Character 23: 't' word=(19, 25): 'thirty'
Character 24: 'y' word=(19, 25): 'thirty'
Character 25: '-' word=(19, 25): 'thirty'
Character 26: 'f' word=(26, 30): 'five'
Character 27: 'i' word=(26, 30): 'five'
Character 28: 'v' word=(26, 30): 'five'
Character 29: 'e' word=(26, 30): 'five'
Character 30: '.' word=(26, 30): 'five'
Character 31: ' ' word=(26, 30): 'five'
Character 32: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 33: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 34: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 35: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 36: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 37: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 38: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 39: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 40: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 41: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 42: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 43: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 44: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 45: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 46: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 47: 'x' word=(32, 48): 'xxxxxxxxxxxxxxxx'
Character 48: 'x' word=(48, 51): 'xxx'
Character 49: 'x' word=(48, 51): 'xxx'
Character 50: 'x' word=(48, 51): 'xxx'
PASS successfullyParsed is true

TEST COMPLETE
